add (broken) new field<> based solver
the idea is to modify the fluid_solver class into something that uses
the field class, to the point where if we want to change the backend to
something other than plain FFTW, the new solver class will not need to
be modified in any way.

the "vorticity_equation" class included in this commit is not
functional, it is only a checkpoint.